pyproject.toml
[project]
name = "jcs"
version = "0.1.0"
description = "JCS AI bot"
requires-python = ">=3.12"

authors = [
    { name = "B.Guna", email = "gunasantosh999@gmail.com" }
]

dependencies = [
    "pytest",
    "ruff",
    "mypy",
    "python-dotenv",
    "openai>=1.75.0",
    "uvicorn>=0.34.2",
    "fastapi[standard]>=0.115.12",
]

